move geoprocessing.json to project dir
