+++ packages/base-project/project/projectClient.ts
@@ -4,8 +4,8 @@ import precalc from "./precalc.json" with { type: "json" };
 import objectives from "./objectives.json" with { type: "json" };
 import geographies from "./geographies.json" with { type: "json" };
 import basic from "./basic.json" with { type: "json" };
+import gp from "./geoprocessing.json" with { type: "json" };
 import projectPackage from "../package.json" with { type: "json" };
-import gp from "../geoprocessing.json" with { type: "json" };
 
 import { ProjectClientBase } from "@seasketch/geoprocessing/client-core";
