+++ packages/geoprocessing/examples/output/polygon/boundaryAreaOverlap.json
@@ -1,0 +1,162 @@
+{
+  "metrics": [
+    {
+      "geographyId": "world",
+      "metricId": "boundaryAreaOverlap",
+      "classId": "eez",
+      "sketchId": "60907815f087a55d051c97ed",
+      "groupId": null,
+      "value": 1842406.504455,
+      "extra": {
+        "sketchName": "test1"
+      }
+    },
+    {
+      "geographyId": "world",
+      "metricId": "boundaryAreaOverlap",
+      "classId": "eez",
+      "sketchId": "609ef1d65b3d7a396cdbc944",
+      "groupId": null,
+      "value": 1933849.993655,
+      "extra": {
+        "sketchName": "polygon",
+        "isCollection": true
+      }
+    },
+    {
+      "geographyId": "world",
+      "metricId": "boundaryAreaOverlap",
+      "classId": "eez",
+      "sketchId": "609ef2525b3d7a396cdbc951",
+      "groupId": null,
+      "value": 0,
+      "extra": {
+        "sketchName": "test2"
+      }
+    },
+    {
+      "geographyId": "world",
+      "metricId": "boundaryAreaOverlap",
+      "classId": "eez",
+      "sketchId": "60ad299c9554743cd643b985",
+      "groupId": null,
+      "value": 91443.489199,
+      "extra": {
+        "sketchName": "test3"
+      }
+    }
+  ],
+  "sketch": {
+    "type": "FeatureCollection",
+    "properties": {
+      "id": "609ef1d65b3d7a396cdbc944",
+      "name": "polygon",
+      "updatedAt": "2021-05-25T16:46:12.125Z",
+      "createdAt": "2021-05-14T21:55:34.020Z",
+      "sketchClassId": "5d0d437fa7c99269b81eff7f",
+      "isCollection": true,
+      "userAttributes": []
+    },
+    "features": [
+      {
+        "type": "Feature",
+        "properties": {
+          "id": "60907815f087a55d051c97ed",
+          "name": "test1",
+          "updatedAt": "2021-05-25T16:46:12.126Z",
+          "createdAt": "2021-05-03T22:24:21.261Z",
+          "sketchClassId": "5d0d435ba7c99269b81eff6c",
+          "isCollection": false,
+          "userAttributes": [
+            {
+              "label": "Type of MMA",
+              "fieldType": "ChoiceField",
+              "exportId": "TYPE_OF_MMA",
+              "value": "Community-Based Fishing Area"
+            },
+            {
+              "label": "What other type?",
+              "fieldType": "TextArea",
+              "exportId": "WHAT_OTHER_TY",
+              "value": ""
+            }
+          ]
+        },
+        "id": "1",
+        "bbox": [
+          -156.92271259686578,
+          21.17195157781254,
+          -156.9017930208691,
+          21.185796391842885
+        ],
+        "geometry": null
+      },
+      {
+        "type": "Feature",
+        "properties": {
+          "id": "609ef2525b3d7a396cdbc951",
+          "name": "test2",
+          "updatedAt": "2021-05-14T21:57:45.842Z",
+          "createdAt": "2021-05-14T21:57:38.768Z",
+          "sketchClassId": "5d0d435ba7c99269b81eff6c",
+          "isCollection": false,
+          "userAttributes": [
+            {
+              "label": "Type of MMA",
+              "fieldType": "ChoiceField",
+              "exportId": "TYPE_OF_MMA",
+              "value": "Marine Life Conservation District"
+            },
+            {
+              "label": "What other type?",
+              "fieldType": "TextArea",
+              "exportId": "WHAT_OTHER_TY",
+              "value": ""
+            }
+          ]
+        },
+        "id": "2",
+        "bbox": [
+          -154.93901851781655,
+          19.532663557180157,
+          -154.78071595699348,
+          19.655159895457032
+        ],
+        "geometry": null
+      },
+      {
+        "type": "Feature",
+        "properties": {
+          "id": "60ad299c9554743cd643b985",
+          "name": "test3",
+          "updatedAt": "2021-05-25T16:45:24.298Z",
+          "createdAt": "2021-05-25T16:45:16.678Z",
+          "sketchClassId": "5d0d435ba7c99269b81eff6c",
+          "isCollection": false,
+          "userAttributes": [
+            {
+              "label": "Type of MMA",
+              "fieldType": "ChoiceField",
+              "exportId": "TYPE_OF_MMA",
+              "value": "Makai Watch"
+            },
+            {
+              "label": "What other type?",
+              "fieldType": "TextArea",
+              "exportId": "WHAT_OTHER_TY",
+              "value": ""
+            }
+          ]
+        },
+        "id": "3",
+        "bbox": [
+          -155.9892625989909,
+          19.8414710982367,
+          -155.9849281480671,
+          19.84550781641468
+        ],
+        "geometry": null
+      }
+    ]
+  }
+}
