+++ packages/geoprocessing/scripts/base/project/projectClient.ts
@@ -24,7 +24,9 @@ export function getProjectClient(projectPath: string) {
   const projectPackage = fs.readJsonSync(
     path.join(projectPath, "package.json")
   );
-  const gp = fs.readJsonSync(path.join(projectPath, "geoprocessing.json"));
+  const gp = fs.readJsonSync(
+    path.join(projectPath, "project", "geoprocessing.json")
+  );
 
   return new ProjectClientBase({
     datasources,
