+++ packages/geoprocessing/scripts/init/createProject.ts
@@ -92,7 +92,7 @@ export async function createProject(
     await $`cp -r ${baseProjectPath}/* ${projectPath}`;
     await $`cp -r ${baseProjectPath}/. ${projectPath}`;
     await $`rm -f ${projectPath}/package-lock.json`;
-    await $`rm -f ${projectPath}/geoprocessing.json`;
+    await $`rm -f ${projectPath}/project/geoprocessing.json`;
     await $`rm -rf ${projectPath}/examples/outputs/*.*`;
     await $`rm -rf ${projectPath}/examples/features/*.*`;
     await $`rm -rf ${projectPath}/examples/sketches/*/*`;
@@ -158,7 +158,7 @@ export async function createProject(
   spinner.start("creating geoprocessing.json");
   const geoAuthor = email ? `${metadata.author} <${email}>` : metadata.author;
   await fs.writeFile(
-    `${projectPath}/geoprocessing.json`,
+    `${projectPath}/project/geoprocessing.json`,
     JSON.stringify(
       {
         author: geoAuthor,
