+++ packages/wiki-docs/Migrating.md
@@ -199,7 +199,7 @@ import objectives from "./objectives.json" with { type: "json" };
 import geographies from "./geographies.json" with { type: "json" };
 import basic from "./basic.json" with { type: "json" };
 import projectPackage from "../package.json" with { type: "json" };
-import gp from "../geoprocessing.json" with { type: "json" };
+import gp from "../project/geoprocessing.json" with { type: "json" };
 
 import { ProjectClientBase } from "@seasketch/geoprocessing/client-core";
 
@@ -295,7 +295,7 @@ import objectives from "./objectives.json" with { type: "json" };
 import geographies from "./geographies.json" with { type: "json" };
 import basic from "./basic.json" with { type: "json" };
 import projectPackage from "../package.json" with { type: "json" };
-import gp from "../geoprocessing.json" with { type: "json" };
+import gp from "../project/geoprocessing.json" with { type: "json" };
 import { ProjectClientBase } from "@seasketch/geoprocessing" with { type: "json" };
 
 const projectClient = new ProjectClientBase({
